backend/ml/air/requirements.txt
certifi==2022.12.7
numpy==1.24.2
packaging==23.0
pandas==2.0.0
patsy==0.5.3
python-dateutil==2.8.2
psycopg2-binary==2.9.6
pytz==2023.3
scipy==1.10.1
six==1.16.0
statsmodels==0.13.5
tzdata==2023.3
